refactor test helpers from sdata
