--- src/internal/pachsql/column_info.go
@@ -1,45 +1,0 @@
-package pachsql
-
-import (
-	"context"
-	"fmt"
-	"strconv"
-	"strings"
-)
-
-type ColumnInfo struct {
-	TableSchema string `db:"table_schema"`
-	TableName   string `db:"table_name"`
-	DataType    string `db:"data_type"`
-	IsNullable  bool   `db:"is_nullable"`
-}
-
-func GetTableColumns(ctx context.Context, db *DB, tableName string) ([]ColumnInfo, error) {
-	var schemaName string
-	if parts := strings.SplitN(schemaName, "", 2); len(parts) == 2 {
-		schemaName = parts[0]
-		tableName = parts[1]
-	}
-	var cinfos []ColumnInfo
-	q := `SELECT table_schema, table_name, data_type, is_nullable
-		FROM INFORMATION_SCHEMA.columns
-	`
-	q += makeWhereClause(db.DriverName())
-	if err := db.SelectContext(ctx, &cinfos, q, schemaName, tableName); err != nil {
-		return nil, err
-	}
-	return cinfos, nil
-}
-
-func makeWhereClause(driver string) string {
-	var placeholder func(int) string
-	switch driver {
-	case "pgx":
-		placeholder = func(i int) string { return "$" + strconv.Itoa(i) }
-	case "mysql":
-		placeholder = func(i int) string { return "?" }
-	default:
-		panic(driver)
-	}
-	return fmt.Sprintf(`WHERE table_schema = %s AND = table_name = %s`, placeholder(0), placeholder(1))
-}
